Completed functionality for superAdmin dashboard
