+++ package.json
@@ -6,12 +6,15 @@
     "@material-ui/core": "^1.4.0",
     "@material-ui/icons": "^1.1.0",
     "axios": "^0.18.0",
+    "chart.js": "^2.7.2",
     "crypto-js": "^3.1.9-1",
     "date-fns": "^2.0.0-alpha.16",
     "flexboxgrid": "^6.3.1",
     "material-ui-pickers": "^1.0.0-rc.14",
     "moment": "^2.22.2",
+    "randomcolor": "^0.5.3",
     "react": "^16.4.1",
+    "react-chartjs-2": "^2.7.4",
     "react-dom": "^16.4.1",
     "react-grecaptcha": "^1.2.5",
     "react-number-format": "^3.5.1",
