+++ src/actions/dashboardAction.js
@@ -1,0 +1,90 @@
+import { hostURL, dashboardAPI } from '../app.config';
+
+var axios = require('axios');
+var randomColor = require('randomcolor');
+
+//Function to get analytics details 
+export function getDashboardDetails(token) {
+
+    var setting = {
+        method: 'post',
+        url: hostURL + dashboardAPI,
+        data: {
+            "platform": "web", 
+	    },
+        headers: {
+            'content-type': 'application/json',
+            'auth' : token
+        }
+    }
+    
+    var response = axios(setting).then(
+        response => 
+        {
+            var responseDetails = response.data.responseData
+
+            var output ={
+                "message": response.data.message,
+                "status":response.data.status,
+                "responseData":{
+                    "openTicketsCount": responseDetails.openTicketsCount,
+                    "openRedeemRequestsCount": responseDetails.openRedeemRequestsCount,
+                    "totalMerchantsCount": responseDetails.totalMerchantsCount,
+                    "activeDealsCount": responseDetails.activeDealsCount, 
+                    "tickets": getFormattedData(responseDetails.tickets), 
+                    "redeemRequests": getFormattedData(responseDetails.redeemRequests) 
+                }
+            }
+
+            return output;
+        }
+        )
+
+        .catch(response => response = {
+            success: 500,
+            message: "Your submission could not be completed. Please Try Again!",
+            responseData: undefined
+        }
+        );
+
+    return {
+        type: 'GET_DASHBOARD_DETAILS',
+        payload: response
+    }
+}
+
+function getFormattedData(dataArray) {
+
+    var arrLabels=[];
+    var arrData=[];
+    var arrBackgroundColor=[];
+    var arrHoverBackground=[];
+
+    if(dataArray !== undefined){
+        dataArray.forEach(element => {
+            arrLabels.push(element.name)
+            arrData.push(element.count)
+        })
+
+        var color = randomColor({
+            count: dataArray.length,
+            luminosity: 'dark',
+            alpha: 0.7,
+            hue: 'random'
+        });
+
+        arrBackgroundColor = color
+        arrHoverBackground = color
+    }
+
+    var dataSet = {};
+    dataSet.data = arrData;
+    dataSet.backgroundColor = arrBackgroundColor;
+    dataSet.hoverBackgroundColor = arrHoverBackground;
+
+    return ({
+        "labels": arrLabels,
+        "datasets":[dataSet]
+    })
+
+}
